Solves for two-letter words using "y" as a vowel
Previous solution failed on line 21, where words like "by" and "my" is made entirely of consonant even though the "y" is being used as a vowel. New commit uses "word[2] != nil" to check if the word is three letters. A two-letter word made up of entirely consonants will either (a) already end or start in a vowel or (b) end with 'y'. Adding the check for a third letter ensures that all two-letter words are covered by the solution. :)